% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_area.R
\name{geom_area}
\alias{geom_area}
\title{Draw an area plot}
\usage{
ggplot(data, aes(x = var)) +
  geom_area()
}
\arguments{
\item{data}{A tibble to be used for the area plot}

\item{var}{Specify which variable should be drawn on the x-axis.}
}
\description{
\code{geom_area} is a geom that should be used in conjunction
with \code{ggplot} to draw an area plot.
}
\examples{
library(tidyverse)
library(gapminder)

# Visualize the distribution of a variable (compare to
# geom_histogram)

ggplot(gapminder, aes(x = lifeExp)) +
  geom_area(stat = "bin")

# Visualize the distribution of a variable by group:

ggplot(gapminder, aes(x = lifeExp, fill = continent)) +
  geom_area(stat = "bin")

# Adjust the previous plot so that groups overlap ("dodge")
# instead of stack:

ggplot(gapminder, aes(x = lifeExp, fill = continent)) +
  geom_area(stat = "bin", pos = "dodge")

# Increase the transparency (alpha) of the area plot
# so that overlapped information is visible:

ggplot(gapminder, aes(x = lifeExp, fill = continent)) +
  geom_area(stat = "bin", pos = "dodge", alpha = .6)

}
