% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_density.R
\name{geom_density}
\alias{geom_density}
\title{Draw a density plot}
\usage{
ggplot(data, aes(x = var)) +
  geom_density()
}
\arguments{
\item{data}{A tibble to be used for the density plot}

\item{var}{Specify which variable should be drawn on the x-axis.}
}
\description{
\code{geom_density} is a geom that should be used in conjunction
with \code{ggplot} to draw a density plot. Compare to histograms or area plots.
}
\examples{
library(tidyverse)
library(gapminder)

# Draw a basic density plot:

ggplot(gapminder, aes(x = lifeExp)) +
  geom_density()

# Compare different groups using color and fill:

gapminder \%>\%
  ggplot(aes(x = lifeExp, color = continent, fill = continent)) +
  geom_density()

# Adjust transparency (alpha):

gapminder \%>\%
 ggplot(aes(x = lifeExp, color = continent, fill = continent)) +
 geom_density(alpha = .3)
}
