% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.r
\name{ggplot}
\alias{ggplot}
\title{Initiate a new ggplot}
\usage{
ggplot(data, aes(x, y, ...))
}
\arguments{
\item{data}{A tibble.}

\item{x}{Specify which variable should be drawn on the x-axis.}

\item{y}{Specify which variable should be drawn on the y-axis.}

\item{...}{Any other aeshetic mappings (color, fill, etc).}
}
\description{
If you declare a dataset and an aesthetic mapping inside the \code{ggplot()} call,
each layer will inherit that information by default.
}
\examples{
library(tidyverse)
library(gapminder)

# Draw a scatterplot:

ggplot(gapminder, aes(x = gdpPercap, y = lifeExp)) +
  geom_point()

# Draw a histogram:

ggplot(gapminder, aes(x = lifeExp)) +
  geom_histogram()

}
