% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.packages.R
\name{install.packages}
\alias{install.packages}
\title{Install a package}
\usage{

}
\description{
To use an add-on package from the web, install it first.
After it's installed, load it into your current R session
using \code{\link[=library]{library()}}.

You only have to install a package once, but you have to
load it using \code{\link[=library]{library()}} every time you open a new R session.
}
\details{
The package name needs quotes.

\code{install.packages(package)}
}
\examples{
install.packages("tidyverse",  dependencies = TRUE)

install.packages("gapminder")

}
\seealso{
\code{\link[=library]{library()}}
}
