% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{library}
\alias{library}
\title{load a package}
\usage{

}
\description{
Load packages you've already installed
into your current R session. After you've
loaded it, all its functions will be
available for you to use.
}
\details{
The package name doesn't need quotes.

\code{library(package)}
}
\examples{

library(tidyverse)

library(gapminder)

}
\seealso{
\code{\link[=install.packages]{install.packages()}}
}
