% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm}
\alias{lm}
\title{Fit a linear model}
\usage{
lm(formula, data)


}
\arguments{
\item{formula}{A symbolic description of the model. For example:
\code{y ~ x} estimates the model \code{y = beta_0 + beta_1*x + u}.}

\item{data}{A data frame like a tibble.}
}
\description{
Run a regression using ordinary least squares.
}
\details{

}
\examples{
# A tibble with the heights and weights
# of 5 individuals:

library(tidyverse)

htwt <- tibble(
  sex = c("male", "male", "female", "female", "male"),
  height = c(62, 72, 68, 64, 70),
  weight = c(169, 243, 147, 135, 183)
)

# Simple regression:

lm(weight ~ height, data = htwt)

# Simple regression with tidied output:

lm(weight ~ height, data = htwt) \%>\%
  broom::tidy(conf.int = TRUE)

-----------------------------------

# Multiple regression:

lm(weight ~ height + sex, data = htwt)

# Multiple regression without an intercept:

lm(weight ~ 0 + height + sex, data = htwt)

-----------------------------------

# Interaction terms: ':' versus '*'
#   Use ':' to include only an interaction.
#   Use '*' to also include the terms alone.
#   So x*z is equivalent to 'x + z + x:z'
#   And `x*y*z` is equivalent to
#     'x + z + y + x:y + x:z + y:z + x:y:z'

lm(weight ~ height:sex, data = htwt)

lm(weight ~ height*sex, data = htwt)

-----------------------------------

# Fitted values and residuals

lm(weight ~ height, data = htwt) \%>\%
  fitted.values()

lm(weight ~ height, data = htwt) \%>\%
  residuals()


}
