% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log}
\alias{log}
\title{Log}
\usage{

}
\description{
Compute the natural logarithm of a value, or the natural log of each element of a vector.
}
\details{
\code{log(x, base = exp(1))}
}
\examples{
# Natural log of e:

log(exp(1))

#> [1] 1

-----------------------------------

# Compute the natural log of every
# element of a vector:

c(1, 2, 3) \%>\% log()

#> [1] 0.0000000 0.6931472 1.0986123

-----------------------------------

# Use log() as a helper function
# with mutate():

library(gapminder)

gapminder \%>\%
  select(country, year, gdpPercap) \%>\%
  mutate(gdp_log = log(gdpPercap))

#> # A tibble: 1,704 x 4
#>   country      year gdpPercap gdp_log
     <fct>       <int>     <dbl>   <dbl>
#> 1 Afghanistan  1952      779.    6.66
#> 2 Afghanistan  1957      821.    6.71
#> 3 Afghanistan  1962      853.    6.75

}
\seealso{
\code{\link[base:Log]{base::log2()}}, \code{\link[base:Log]{base::log10()}}, \code{\link[base:Log]{base::exp()}}, \code{\link[=mutate]{mutate()}}
}
