% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\title{Arithmetic mean}
\usage{

}
\description{
Calculates the arithmetic mean of a numeric vector.
}
\details{
The first argument \code{x} is a numeric vector. Setting
\code{na.rm = TRUE} will remove \code{NA}'s in the vector before
calculating the mean.

\code{mean(x, na.rm = FALSE)}
}
\examples{
c(1, 2, 3) \%>\% mean()
#> [1] 2

-----------------------------------

# Missing values

c(1, 2, NA) \%>\% mean()
#> [1] NA

c(1, 2, NA) \%>\% mean(na.rm = T)
#> [1] 1.5

-----------------------------------

library(gapminder)

# Average life expectancy of countries
# each year:

gapminder \%>\%
  group_by(year) \%>\%
  summarize(lifeExp_mean = mean(lifeExp))

#> A tibble: 12 x 2
#>    year lifeExp_mean
     <int>        <dbl>
#> 1  1952         49.1
#> 2  1957         51.5
#> 3  1962         53.6

}
\seealso{
\code{\link[=min]{min()}}, \code{\link[=max]{max()}}, \code{\link[=median]{median()}}, \code{\link[=quantile]{quantile()}}
}
