% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{Mutate a tibble by adding variables}
\usage{

}
\description{
Adds new variables that may be combinations or transformations of existing variables.
}
\details{
The first argument \code{\link[=mutate]{mutate()}} takes is a tibble.
After that, you'll name the new variables you'd
like to create and how you want to define them.

\code{mutate(data, ...)}

\code{data \%>\% mutate(...)}
}
\examples{
# Define a new variable that's a
# combination of other variables:

tibble(x = c(1, 1, 2),
       y = c(1, 3, 0)) \%>\%
  mutate(z = 2*x + y)

#> # A tibble: 4 x 3
#>    x     y     z
  <dbl> <dbl> <dbl>
#>    1     1     3
#>    1     3     5
#>    2     0     4

-----------------------------------

library(gapminder)

gapminder \%>\%
  mutate(gdp_total = gdpPercap*pop) \%>\%
  select(-continent, -lifeExp)

#> # A tibble: 1,704 x 5
#>   country      year      pop gdpPercap    gdp_total
     <fct>       <int>    <int>     <dbl>        <dbl>
#> 1 Afghanistan  1952  8425333      779.  6567086330.
#> 2 Afghanistan  1957  9240934      821.  7585448670.
#> 3 Afghanistan  1962 10267083      853.  8758855797.

}
\seealso{
Other dplyr verbs: \code{\link[=filter]{filter()}}, \code{\link[=group_by]{group_by()}}, \code{\link[=arrange]{arrange()}}, \code{\link[=summarize]{summarize()}}, \code{\link[=select]{select()}}
}
