% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\name{nrow}
\alias{nrow}
\title{Number of rows in a data frame}
\usage{

}
\description{
Since rows are also observations, this is also the
number of observations in a data frame.
}
\details{
\code{nrow(data)}
}
\examples{
tibble(
  x = 1:5,
  y = 6:10
) \%>\%
  nrow()

#> [1] 5

-----------------------------------

library(gapminder)

gapminder \%>\% nrow()

#> [1] 1704

-----------------------------------

# Use nrow with filter:

gapminder \%>\%
  filter(lifeExp > 80) \%>\%
  nrow()

#> [1] 21

}
\seealso{
\code{\link[=ncol]{ncol()}}, \code{\link[base:names]{base::names()}}
}
