% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot.R
\name{qplot}
\alias{qplot}
\title{Draw a quick plot}
\usage{
qplot(x, y, ..., data, facets, geom, main)


}
\arguments{
\item{x}{Specify which variable should be drawn on the x-axis.}

\item{y}{Specify which variable should be drawn on the y-axis.
If you're drawing the distribution of a single variable with
a histogram or similar geom, specifying both x and y may be
unnecessary.}

\item{...}{Optionally include which variables should be represented
by \code{color}, \code{fill}, \code{shape}, and \code{size}. \code{color} refers to
the color of the geom outline, and \code{fill} refers to the color of
the interior.}

\item{data}{A data frame like a tibble.}

\item{facets}{A faceting formula. For example, to facet by the
variable \code{z}, \code{facets = ~ z}}

\item{geom}{A character vector of geometries to draw. Any geom
from ggplot will work. For example, to use \code{geom_boxplot},
use \code{geom = "boxplot"}.}

\item{main}{A character string for the main plot title.}
}
\description{
Use \code{qplot()} to draw simple ggplots. For example, you can draw a histogram
to visualize the distribution of a single variable (\code{geom = "histogram"}).
You can draw a scatterplot to visualize the relationship between two
variables (\code{geom = "point"}) and even add a line of best fit
(\code{geom = c("point", "smooth")}). You can also represent a variable using
a different \code{color} (and \code{fill}, \code{shape}, \code{size}). Finally, you can \code{facet} by a
discrete variable to draw multiple plots for each realization of that
variable.
}
\details{

}
\examples{
# A tibble with the heights and weights of 4 individuals:

htwt <- tibble(
  sex = c("male", "male", "female", "female"),
  height = c(62, 72, 68, 64),
  weight = c(169, 243, 147, 135)
)

# A simple scatterplot with sex represented by `color`:

qplot(
  data = htwt,
  x = height,
  y = weight,
  color = sex,
  geom = "point",
  main = "Heights and Weights of 4 Individuals"
  )

-----------------------------------

# A simple histogram with sex represented by `fill`:

qplot(
  data = htwt,
  x = height,
  fill = sex,
  geom = "histogram",
  bins = 4,
  main = "Frequency of Height Among 4 Individuals"
  )

-----------------------------------

# Faceting by sex:

qplot(
  data = htwt,
  x = height,
  y = weight,
  color = sex,
  geom = c("point", "line"),
  facets = ~ sex,
  main = "Heights and Weights of 4 Individuals"
  )

}
\seealso{
\code{\link[=ggplot]{ggplot()}}, \code{\link[=geom_histogram]{geom_histogram()}}, \code{\link[=geom_point]{geom_point()}}, \code{\link[=geom_smooth]{geom_smooth()}}
}
