% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile}
\alias{quantile}
\title{Quantiles}
\usage{

}
\description{
Calculates the quantiles of a numeric vector.
}
\details{
The first argument \code{x} is a numeric vector.

The second argument is a vector of probabilities.
For instance, \code{probs = .5} will return the median of \code{x}.

\code{quantile(x, probs)}
}
\examples{
1:10 \%>\%
  quantile(probs = c(.25, .5, .75))

#>  25\%  50\%  75\%
#> 3.25 5.50 7.75

-----------------------------------

library(gapminder)

gapminder \%>\%
  summarize(pops = quantile(pop, c(.1, .5, .9)))

#>          pops
           <dbl>
#> 1     946367.
#> 2    7023596.
#> 3   54801370.

}
\seealso{
\code{\link[=median]{median()}}, \code{\link[=mean]{mean()}}, \code{\link[=min]{min()}}, \code{\link[=max]{max()}}, \code{\link[=sd]{sd()}}
}
