% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce}
\alias{reduce}
\title{Reduce a vector to a single value by iteratively applying a binary function}
\usage{
reduce(.x, .f)
}
\arguments{
\item{.x}{A list or vector}

\item{.f}{A binary (2 argument function), which will be iteratively applied to the vector \code{.x}}
}
\description{
Reduce a vector, \code{.x}, to a single value by calling a binary
function (a function that takes two values), \code{.f}. \code{reduce} works by first applying
the function to the first two elements of \code{.x}, then calling the function on the
result of the first computation and the third element of \code{.x}, then on the result of that and the fourth element, and so on. Calling \code{reduce}
on a vector with three elements using the function \code{f} is equivalent to calling \code{f(f(x1, x2), x3)}.
}
\examples{
# `sum` is just reduce with .f = `+`:

reduce(c(1, 2, 3), `+`)

#> 6

-----------------------------------

# .f = intersect

reduce(
  list(c(2, 6, 3, 4), c(4, 3, 2), c(1, 2, 3)),
  intersect
  )

#> 2 3

-----------------------------------

# .f = `*`

1:10 \%>\% reduce(`*`)

#> 3628800
}
