% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.R
\name{rep}
\alias{rep}
\title{Repeat values}
\usage{

}
\description{
Create a vector of repeated elements.
}
\details{
First argument \code{x}: a value or a vector. It can be any data type.

Second argument \code{times}: the number of times you'd like to repeat \code{x}.

Third argument \code{each}: if \code{x} is a vector, \code{each} will be the number
of times you'd like to repeat each element of \code{x}.

\code{rep(x, times = 1, each = 1)}
}
\examples{
rep(1, times = 3)

#> [1] 1 1 1

-----------------------------------

rep(1:5, each = 2)

#> [1] 1 1 2 2 3 3 4 4 5 5

-----------------------------------

c(rep("a", times = 3), rep("b", times = 2))

#> [1] "a" "a" "a" "b" "b"

}
\seealso{
\code{\link[=seq]{seq()}}, \code{\link[=c]{c()}}, \code{\link[=sample]{sample()}}
}
