% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Random samples}
\usage{

}
\description{
Draw random samples from a vector of elements.
}
\details{
First argument \code{x}: a vector. Can be any data type.

Second argument \code{size}: the length of the output vector,
that is, the number of random draws you'd like to make.

Third argument \code{replace}: by default, replace is set to
\code{FALSE}, so sampling is done without replacement. If you'd
like to sample \emph{with} replacement, set \code{replace = TRUE}.
If you've set \code{size} to be greater than the length of x,
you'll need to set \code{replace} to be TRUE.

Fourth argument \code{prob}: a vector of weights which you
should define if you don't want there to be an equal
probability for drawing each element of x.

\code{sample(x, size, replace = FALSE, prob = NULL)}
}
\examples{
# Flip a coin 10 times:

sample(c("heads", "tails"),
       size = 4, replace = TRUE)

#> [1] "tails" "tails" "tails" "heads"

-----------------------------------

sample(1:3, size = 12, replace = TRUE,
       prob = c(.5, .3, .2))

#> [1] 2 3 2 1 1 1 2 1 1 2 3 2

}
\seealso{
\code{\link[=seq]{seq()}}, \code{\link[=rep]{rep()}}, \code{\link[=c]{c()}}, \code{\link[stats:Uniform]{stats::runif()}}, \code{\link[stats:Normal]{stats::rnorm()}}
}
