% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd}
\alias{sd}
\title{Standard deviation}
\usage{

}
\description{
Returns the standard deviation of the elements of x.
}
\details{
First argument \code{x}: is a numeric vector.

Second argument \code{na.rm = FALSE}: by default, NA's are
left in the data. Setting \code{na.rm = TRUE} will remove
NA's in the vector before calculating the standard deviation.

\code{sd(x, na.rm = FALSE)}
}
\examples{
c(4, 6, 2, 5) \%>\% sd()

#> [1] 1.707825

-----------------------------------

library(gapminder)

gapminder \%>\%
  filter(year == 2007) \%>\%
  group_by(continent) \%>\%
  summarize(lifeExp_sd = sd(lifeExp))

#> # A tibble: 5 x 2
#>   continent lifeExp_sd
     <fct>          <dbl>
#> 1 Africa         9.63
#> 2 Americas       4.44
#> 3 Asia           7.96
#> 4 Europe         2.98
#> 5 Oceania        0.729

}
\seealso{
\code{\link[=mean]{mean()}}, \code{\link[=min]{min()}}, \code{\link[=max]{max()}}, \code{\link[=quantile]{quantile()}}, \code{\link[=var]{var()}}
}
