% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{arrange}
\alias{arrange}
\title{Arrange rows of a tibble}
\usage{
arrange(data, ...)
}
\description{
Orders rows of a tibble by the variable of your choice.
}
\examples{
msleep \%>\% arrange(sleep_total)

msleep \%>\% arrange(desc(sleep_total))

}
\seealso{
select

group_by

summarize

mutate

filter
}
